# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Multi-channel communications gateway bridging Meshtastic, Iridium, Cellular, APRS, TAK, MQTT, ZigBee, and Webhooks.",
        "title": "MeshSat API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6050",
    "basePath": "/",
    "paths": {
...
        "/api/messages": {
            "get": {
                "description": "Returns paginated mesh messages with optional filters",
                "tags": [
                    "messages"
                ],
                "summary": "Get message history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by node ID (!hex format)",
                        "name": "node",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339)",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339)",
                        "name": "until",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by port number",
                        "name": "portnum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport (radio, mqtt, satellite)",
                        "name": "transport",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by direction (rx, tx)",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Results per page (default 50, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "messages, total, limit, offset",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/messages/send": {
            "post": {
                "description": "Sends a text message through the Meshtastic radio",
                "tags": [
                    "messages"
                ],
                "summary": "Send a mesh message",
                "parameters": [
                    {
                        "description": "Message to send",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.SendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "transport.SendRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
    }
}